+++ src/webview/components/NumberField.tsx
@@ -1,4 +1,5 @@
 import * as React from 'react';
+import { useMemo } from 'react';
 
 export interface INumberFieldProps {
   label: string;
@@ -7,11 +8,22 @@ export interface INumberFieldProps {
   onChange: (value: number) => void;
   isFloat?: boolean;
   isDisabled?: boolean;
+  isRange?: boolean;
 
   [prop: string]: any; // Allow any other property
 }
 
-export const NumberField: React.FunctionComponent<INumberFieldProps> = ({ label, placeholder, value, onChange, isFloat, isDisabled, ...props }: React.PropsWithChildren<INumberFieldProps>) => {
+export const NumberField: React.FunctionComponent<INumberFieldProps> = ({ label, placeholder, value, onChange, isFloat, 
+  isRange, isDisabled, ...props }: React.PropsWithChildren<INumberFieldProps>) => {
+
+  const styles = useMemo(() => {
+    if (isRange) {
+      return `mt-4 h-2 rounded-lg appearance-none cursor-pointer focus:outline-none focus:ring-0 bg-[var(--vscode-checkbox-background)] border border-[var(--vscode-checkbox-border)]`;
+    }
+
+    return `mt-1 block p-2 rounded-md border-[var(--vscode-panel-border)] bg-[var(--vscode-input-background)] text-[var(--vscode-input-foreground)] shadow-sm sm:text-sm focus:outline-[var(--vscode-focusBorder)] focus:outline-1 disabled:opacity-75`
+  }, [isRange])
+
   return (
     <div className='w-full'>
       <label htmlFor={label.replace(' ', '-').toLowerCase()} className="block text-sm font-medium text-[var(--vscode-editor-foreground)]">
@@ -19,10 +31,10 @@ export const NumberField: React.FunctionComponent<INumberFieldProps> = ({ label,
       </label>
       <div className="mt-1">
         <input
-          type="number"
+          type={isRange ? 'range' : 'number'}
           name={label.replace(' ', '-').toLowerCase()}
           value={value}
-          className="mt-1 block w-full p-2 rounded-md border-[var(--vscode-panel-border)] bg-[var(--vscode-input-background)] text-[var(--vscode-input-foreground)] shadow-sm sm:text-sm focus:outline-[var(--vscode-focusBorder)] focus:outline-1 disabled:opacity-75"
+          className={`w-full ${styles}`}
           placeholder={placeholder}
           disabled={isDisabled}
           onChange={(e) => {