+++ src/webview/constants/Defaults.ts
@@ -1,4 +1,8 @@
 export const Defaults = {
+  fontSize: 16,
+  innerWidth: 100,
+  innerPadding: 2,
+  innerBorder: 15,
   width: 1200,
   height: 675
 }