+++ src/webview/constants/Gradients.ts
@@ -2,5 +2,6 @@ export const Gradients = [
   "radial-gradient(circle, rgba(63,94,251,1) 0%, rgba(252,70,107,1) 100%)",
   "radial-gradient(circle, rgba(238,174,202,1) 0%, rgba(148,187,233,1) 100%)",
   "linear-gradient(to right top, #051937, #004d7a, #008793, #00bf72, #a8eb12)",
-  "linear-gradient(to right bottom, #f141a8, #ff6180, #ff9460, #ffc855, #f9f871)"
+  "linear-gradient(to right bottom, #f141a8, #ff6180, #ff9460, #ffc855, #f9f871)",
+  "transparent",
 ];