+++ src/webview/state/ScreenshotDetailsState.ts
@@ -1,3 +1,4 @@
+import { Defaults } from './../constants/Defaults';
 import { atom } from "recoil";
 import { ScreenshotDetails } from '../models'
 
@@ -5,10 +6,11 @@ export const ScreenshotDetailsState = atom<ScreenshotDetails>({
   key: 'ScreenshotDetailsState',
   default: {
     preset: 'Twitter',
-    fontSize: 13,
+    fontSize: Defaults.fontSize,
     fontFamily: "editor",
-    innerWidth: 100,
-    innerPadding: 2,
+    innerWidth: Defaults.innerWidth,
+    innerPadding: Defaults.innerPadding,
+    innerBorder: Defaults.innerBorder,
     linkColor: undefined,
     width: undefined,
     height: undefined,