+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## [0.0.6] 2023-02-19
+
+- Changed `number` to `range` inputs for width, padding, and font size
+- Added a new border radius option
+- Added transparent to the background color options
+- New preset for Open Graph image (1200x630)
+- Copy to clipboard option
+
 ## [0.0.5] 2023-02-18
 
 - Code refactoring